Add `c` script to open VSCode 🎉
Because I have to use the terminals secure entry option, VSCodes
built-in `code` command doesn’t behave nicely. It opens but doesn’t
focus, so I always have to Tab in manually. Annooooyiiiiiing 🤬

This new little script just uses `open -a "Visual Studio Code"` instead,
which always brings the app to the front. Even better! It searches the
target directory for a `.code-workspace` file and opens that if found,
falling back to the directory itself otherwise.

Basically, it’s similar to `code` but it actually works for me 🤷🏻‍♀️